Optimize GameState.IsGameOver().
This doesn't really have a noticable effect on the performance of the arbiter.
I just did it out of principle.

The idea is that a game is over if and only if there is no 2x6 rectangular area
in the grid with at most 4 occupied cells (the maximum allowed overlap when
placing a new tile). Note that if there are multiple such areas, not all of them
may be valid placements because some do not connect to existing tiles, but if
there is at least one empty area that implies there must be an empty area
adjacent to an occupied cell too.

That means we can generate all possible rectangles once (which _GenerateMoves()
does lazily) and then keep track of the first valid rectangle as we update the
grid. There are about O(HEIGHT * WIDTH * 2) possible moves, and checking each
move takes O(2 * COLORS) time, so the overall complexity becomes:

   O((HEIGHT * WIDTH * 2 + moves) * (2 * COLORS))

where moves is the number of moves in the complete game, which is proportional
to HEIGHT*WIDTH, so the whole thing boils down to O(HEIGHT * WIDTH * COLORS)
with relatively low constants.

Note that this is per game. So the complexity of placing each move becomes
O(COLORS) amortized.