Implement greedy solution based on 1-ply lookahead and a simple evaluation function
The idea is to maximize my own score, while minimizing the score of the
next-best color, where a score is calculated based on recognizing the
following patterns:

  1. one cell of one color
  2. two cells of one color in the same row or column
  3. three cells of one color in the same row and column
  4. four cells of one color in the same row and column (i.e., a square!)

Of course the squares is what we really care about, but the other metrics
are useful to see if we are building towards a square or not.

This approach seems to be able to beat CAIA player3 by a small margin.