Make first line of stderr consistent with Caia `competition` binary.
Outputting R as the first character signals that this program is
random, so it should play multiple matches against each opponent,
even when that opponent plays deterministically.

Following this convention is not really important for me since I
don't use caiaio to test (and the official CodeCup competition
doesn't use this convention) but it seems nice to do it just for
being consistent with the sample players.