Tweak score function once again.
Edges should be double counted after all, because they can participate in the
creation of squares in multiple ways, for example:

   a.b.c
   .....
   c.d.e

If b and d are of my color, I can complete a square a-b-c-d or b-c-e-d.
And either one or both of these can be blocked by fixed digits (or by the
edge of the board). So I should not assume it's sufficient to only count
the edge b-d as part of one of the two squares.